Keep workflow dashboard helpers easier to reuse
The WorkflowsTab component mixed pure node-style lookup, workflow filtering, and schema example generation directly into the view, so this extracts those helpers into a sibling module and adds focused coverage without changing dashboard behavior.

Constraint: Refactor must preserve workflow dashboard rendering and generated example input semantics exactly
Constraint: veryfront-code pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Split the detail rendering subtrees into extra components | would widen this pass beyond the narrow pure-helper extraction targeted by the wave
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep pure workflow dashboard helpers in the helper module unless a future change needs direct access to component-local state
Tested: deno test --no-check --allow-all src/server/dev-ui/dashboard/components/WorkflowsTab.helpers.test.ts; deno check src/server/dev-ui/dashboard/components/WorkflowsTab.tsx src/server/dev-ui/dashboard/components/WorkflowsTab.helpers.ts src/server/dev-ui/dashboard/components/WorkflowsTab.helpers.test.ts; deno fmt --check src/server/dev-ui/dashboard/components/WorkflowsTab.tsx src/server/dev-ui/dashboard/components/WorkflowsTab.helpers.ts src/server/dev-ui/dashboard/components/WorkflowsTab.helpers.test.ts; deno lint src/server/dev-ui/dashboard/components/WorkflowsTab.tsx src/server/dev-ui/dashboard/components/WorkflowsTab.helpers.ts src/server/dev-ui/dashboard/components/WorkflowsTab.helpers.test.ts; git diff --check
Not-tested: Full veryfront-code repo suite
